plantit updates
