--- Singularity
@@ -1,82 +1,0 @@
-BootStrap: docker
-From: ubuntu:18.04
-
-%labels
-  Maintainer: Suxing Liu
-  Version v1.01
-
-%setup
-  #----------------------------------------------------------------------
-  # commands to be executed on host outside container during bootstrap
-  #----------------------------------------------------------------------
-  mkdir ${SINGULARITY_ROOTFS}/opt/code/
-
-%files
-  ./* /opt/code/
-
-%post
-  #----------------------------------------------------------
-  # Install common dependencies and create default entrypoint,
-  # commands to be executed inside container during bootstrap
-  #----------------------------------------------------------
-  # Install dependencies
-  apt update
-  apt install -y \
-    build-essential \
-    python3-setuptools \
-    python3-pip \
-    python3-numexpr \
-    libgl1-mesa-glx \
-    libsm6 \
-    libxext6 \
-    libfontconfig1 \
-    libxrender1 \
-    python3-pandas
-    
-  pip3 install numpy \
-                Pillow \
-                scipy \
-                scikit-image \
-                matplotlib \
-                opencv-python \
-                openpyxl \
-                python-magic \
-                imutils \
-                tabulate \
-                skan \
-                networkx \
-                pandas \
-                numba \
-                psutil
-
-  pip3 install -U scikit-learn
-  
-
-  mkdir /lscratch /db /work /scratch
-  
-  chmod -R a+rwx /opt/code/
-  
-%environment
-  #----------------------------------------------------------
-  # Setup environment variables
-  #----------------------------------------------------------
-  PYTHONPATH=$PYTHONPATH:/opt/code/
-  export PATH
-  LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/opt/code/
-  export LD_LIBRARY_PATH
-
-%runscript
-  #----------------------------------------------------------
-  # Run scripts inside container
-  #----------------------------------------------------------
-   # commands to be executed when the container runs
-   echo "Arguments received: $*"
-   exec /usr/bin/python "$@"
-  
-%test
-  #----------------------------------------------------------
-  # commands to be executed within container at close of bootstrap process
-  #----------------------------------------------------------
-   python3 --version
-   #python3 requirement.py 
-
