--- core/cli.py
@@ -1,1055 +1,0 @@
-import os
-from contextlib import closing
-from multiprocessing import cpu_count, Pool
-from os.path import join, abspath, splitext, getsize, basename, dirname
-from pathlib import Path
-import csv
-import math
-from collections import Counter
-
-import imutils
-import click
-import cv2
-import numpy as np
-from scipy import optimize, ndimage
-from scipy.interpolate import interp1d
-from scipy.spatial import distance as dist
-from skimage import img_as_ubyte, img_as_float, img_as_bool, morphology
-from skimage.color import rgb2lab
-from skimage.feature import peak_local_max
-from skimage.segmentation import clear_border, watershed
-import openpyxl
-from sklearn.cluster import KMeans
-import matplotlib.pyplot as plt
-
-
-@click.group()
-def cli():
-    pass
-
-
-def comp_external_contour_1(orig, thresh, output_directory):
-    # find contours and get the external one
-    contours, hier = cv2.findContours(thresh, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-
-    img_height, img_width, img_channels = orig.shape
-
-    index = 1
-
-    print("contour length {}".format(len(contours)))
-
-    list_of_pts = []
-
-    if len(contours) > 1:
-        '''
-        for ctr in contours:
-
-            list_of_pts += [pt[0] for pt in ctr]
-
-        center_pt = np.array(list_of_pts).mean(axis = 0) # get origin
-
-        clock_ang_dist = clockwise_angle_and_distance(center_pt) # set origin
-
-        list_of_pts = sorted(list_of_pts, key=clock_ang_dist) # use to sort
-
-        contours_joined = np.array(list_of_pts).reshape((-1,1,2)).astype(np.int32)
-
-
-        '''
-        kernel = np.ones((4, 4), np.uint8)
-
-        dilation = cv2.dilate(thresh.copy(), kernel, iterations=1)
-
-        closing = cv2.morphologyEx(dilation, cv2.MORPH_CLOSE, kernel)
-
-        trait_img = closing
-
-    # trait_img = cv2.drawContours(thresh, contours_joined, -1, (0,255,255), -1)
-
-    # x, y, w, h = cv2.boundingRect(contours_joined)
-
-    # trait_img = cv2.rectangle(thresh, (x, y), (x+w, y+h), (255, 255, 0), 3)
-
-    contours, hier = cv2.findContours(trait_img, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-
-    print("contour length {}".format(len(contours)))
-
-    for c in contours:
-
-        # get the bounding rect
-        x, y, w, h = cv2.boundingRect(c)
-
-        # if w>img_width*0.05 and h>img_height*0.05:
-
-        if w > 0 and h > 0:
-            offset_w = int(w * 0.05)
-            offset_h = int(h * 0.05)
-            # draw a green rectangle to visualize the bounding rect
-            roi = orig[y - offset_h: y + h + offset_h, x - offset_w: x + w + offset_w]
-
-            print("ROI {} detected ...".format(index))
-
-            cv2.imwrite(join(output_directory, str(format(index, "02")) + '.png'), roi)
-
-            trait_img = cv2.rectangle(orig, (x, y), (x + w, y + h), (255, 255, 0), 3)
-
-            # trait_img = cv2.putText(orig, "#{}".format(index), (int(x), int(y)), cv2.FONT_HERSHEY_SIMPLEX, 3.0, (255, 0, 255), 10)
-
-            index += 1
-
-    return trait_img
-
-
-def color_cluster_seg(image, args_colorspace, args_channels, args_num_clusters, min_size):
-    # Change image color space, if necessary.
-    colorSpace = args_colorspace.lower()
-
-    if colorSpace == 'hsv':
-        image = cv2.cvtColor(image, cv2.COLOR_BGR2HSV)
-
-    elif colorSpace == 'ycrcb' or colorSpace == 'ycc':
-        image = cv2.cvtColor(image, cv2.COLOR_BGR2YCrCb)
-
-    elif colorSpace == 'lab':
-        image = cv2.cvtColor(image, cv2.COLOR_BGR2LAB)
-
-    else:
-        colorSpace = 'bgr'  # set for file naming purposes
-
-    # Keep only the selected channels for K-means clustering.
-    if args_channels != 'all':
-        channels = cv2.split(image)
-        channelIndices = []
-        for char in args_channels:
-            channelIndices.append(int(char))
-        image = image[:, :, channelIndices]
-        if len(image.shape) == 2:
-            image.reshape(image.shape[0], image.shape[1], 1)
-
-    (width, height, n_channel) = image.shape
-
-    # print("image shape: \n")
-    # print(width, height, n_channel)
-
-    # Flatten the 2D image array into an MxN feature vector, where M is the number of pixels and N is the dimension (number of channels).
-    reshaped = image.reshape(image.shape[0] * image.shape[1], image.shape[2])
-
-    # Perform K-means clustering.
-    if args_num_clusters < 2:
-        print('Warning: num-clusters < 2 invalid. Using num-clusters = 2')
-
-    # define number of cluster
-    numClusters = max(2, args_num_clusters)
-
-    # clustering method
-    kmeans = KMeans(n_clusters=numClusters, n_init=40, max_iter=500).fit(reshaped)
-
-    # get lables
-    pred_label = kmeans.labels_
-
-    # Reshape result back into a 2D array, where each element represents the corresponding pixel's cluster index (0 to K - 1).
-    clustering = np.reshape(np.array(pred_label, dtype=np.uint8), (image.shape[0], image.shape[1]))
-
-    # Sort the cluster labels in order of the frequency with which they occur.
-    sortedLabels = sorted([n for n in range(numClusters)], key=lambda x: -np.sum(clustering == x))
-
-    # Initialize K-means grayscale image; set pixel colors based on clustering.
-    kmeansImage = np.zeros(image.shape[:2], dtype=np.uint8)
-    for i, label in enumerate(sortedLabels):
-        kmeansImage[clustering == label] = int(255 / (numClusters - 1)) * i
-
-    ret, thresh = cv2.threshold(kmeansImage, 0, 255, cv2.THRESH_BINARY | cv2.THRESH_OTSU)
-
-    thresh_cleaned = clear_border(thresh)
-
-    if np.count_nonzero(thresh) > 0:
-
-        thresh_cleaned_bw = clear_border(thresh)
-    else:
-        thresh_cleaned_bw = thresh
-
-    nb_components, output, stats, centroids = cv2.connectedComponentsWithStats(thresh_cleaned, connectivity=8)
-
-    # stats[0], centroids[0] are for the background label. ignore
-    # cv2.CC_STAT_LEFT, cv2.CC_STAT_TOP, cv2.CC_STAT_WIDTH, cv2.CC_STAT_HEIGHT
-    sizes = stats[1:, cv2.CC_STAT_AREA]
-
-    Coord_left = stats[1:, cv2.CC_STAT_LEFT]
-
-    Coord_top = stats[1:, cv2.CC_STAT_TOP]
-
-    Coord_width = stats[1:, cv2.CC_STAT_WIDTH]
-
-    Coord_height = stats[1:, cv2.CC_STAT_HEIGHT]
-
-    Coord_centroids = centroids
-
-    # print("Coord_centroids {}\n".format(centroids[1][1]))
-
-    # print("[width, height] {} {}\n".format(width, height))
-
-    nb_components = nb_components - 1
-
-    # min_size = 70
-
-    max_size = width * height * 0.1
-
-    img_thresh = np.zeros([width, height], dtype=np.uint8)
-
-    # for every component in the image, keep it only if it's above min_size
-    for i in range(0, nb_components):
-        '''
-        #print("{} nb_components found".format(i))
-
-        if (sizes[i] >= min_size) and (Coord_left[i] > 1) and (Coord_top[i] > 1) and (Coord_width[i] - Coord_left[i] > 0) and (Coord_height[i] - Coord_top[i] > 0) and (centroids[i][0] - width*0.5 < 10) and ((centroids[i][1] - height*0.5 < 10)) and ((sizes[i] <= max_size)):
-            img_thresh[output == i + 1] = 255
-
-            print("Foreground center found ")
-
-        elif ((Coord_width[i] - Coord_left[i])*0.5 - width < 15) and (centroids[i][0] - width*0.5 < 15) and (centroids[i][1] - height*0.5 < 15) and ((sizes[i] <= max_size)):
-            imax = max(enumerate(sizes), key=(lambda x: x[1]))[0] + 1    
-            img_thresh[output == imax] = 255
-            print("Foreground max found ")
-        '''
-
-        if (sizes[i] >= min_size):
-            img_thresh[output == i + 1] = 255
-
-    # from skimage import img_as_ubyte
-
-    # img_thresh = img_as_ubyte(img_thresh)
-
-    # print("img_thresh.dtype")
-    # print(img_thresh.dtype)
-
-    # return img_thresh
-    return img_thresh
-
-
-def segmentation(image_file, output_directory):
-    file_name, _ = splitext(image_file)
-    file_size = getsize(image_file) / float(1 << 20)
-
-    print("Segmenting image {0} \n".format(str(file_name)))
-    if file_size > 5.0: print("It will take some time due to large file size {0} MB".format(str(int(file_size))))
-
-    image = cv2.imread(image_file)
-    image_copy = image.copy()
-
-    # color clustering based plant object segmentation
-    masked_image = color_cluster_seg(image_copy, 'lab', '1', 2, 100)
-    cv2.imwrite(join(output_directory, f"{file_name}.mask.png"), masked_image)
-
-    # find external contour and segment image into small ROI based on each plant
-    contours_image = comp_external_contour_1(image.copy(), masked_image, output_directory)
-    cv2.imwrite(join(output_directory, f"{file_name}.label.png"), contours_image)
-
-    return masked_image
-
-
-def comp_external_contour_2(orig, thresh):
-    # find contours and get the external one
-    contours, hier = cv2.findContours(thresh, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-
-    img_height, img_width, img_channels = orig.shape
-
-    index = 1
-
-    for c in contours:
-
-        # get the bounding rect
-        x, y, w, h = cv2.boundingRect(c)
-
-        if w > img_width * 0.01 and h > img_height * 0.01:
-
-            trait_img = cv2.drawContours(orig, contours, -1, (255, 255, 0), 1)
-
-            # draw a green rectangle to visualize the bounding rect
-            roi = orig[y:y + h, x:x + w]
-
-            print("ROI {} detected ...\n".format(index))
-            # result_file = (save_path +  str(index) + file_extension)
-            # cv2.imwrite(result_file, roi)
-
-            trait_img = cv2.rectangle(orig, (x, y), (x + w, y + h), (255, 255, 0), 3)
-
-            index += 1
-
-            '''
-            #get the min area rect
-            rect = cv2.minAreaRect(c)
-            box = cv2.boxPoints(rect)
-            # convert all coordinates floating point values to int
-            box = np.int0(box)
-            #draw a red 'nghien' rectangle
-            trait_img = cv2.drawContours(orig, [box], 0, (0, 0, 255))
-            '''
-            # get convex hull
-            hull = cv2.convexHull(c)
-            # draw it in red color
-            trait_img = cv2.drawContours(orig, [hull], -1, (0, 0, 255), 3)
-
-            '''
-            # calculate epsilon base on contour's perimeter
-            # contour's perimeter is returned by cv2.arcLength
-            epsilon = 0.01 * cv2.arcLength(c, True)
-            # get approx polygons
-            approx = cv2.approxPolyDP(c, epsilon, True)
-            # draw approx polygons
-            trait_img = cv2.drawContours(orig, [approx], -1, (0, 255, 0), 1)
-
-            # hull is convex shape as a polygon
-            hull = cv2.convexHull(c)
-            trait_img = cv2.drawContours(orig, [hull], -1, (0, 0, 255))
-            '''
-
-            '''
-            #get the min enclosing circle
-            (x, y), radius = cv2.minEnclosingCircle(c)
-            # convert all values to int
-            center = (int(x), int(y))
-            radius = int(radius)
-            # and draw the circle in blue
-            trait_img = cv2.circle(orig, center, radius, (255, 0, 0), 2)
-            '''
-
-            area = cv2.contourArea(c)
-            print("Leaf area = {0:.2f}... \n".format(area))
-
-            hull = cv2.convexHull(c)
-            hull_area = cv2.contourArea(hull)
-            solidity = float(area) / hull_area
-            print("solidity = {0:.2f}... \n".format(solidity))
-
-            extLeft = tuple(c[c[:, :, 0].argmin()][0])
-            extRight = tuple(c[c[:, :, 0].argmax()][0])
-            extTop = tuple(c[c[:, :, 1].argmin()][0])
-            extBot = tuple(c[c[:, :, 1].argmax()][0])
-
-            trait_img = cv2.circle(orig, extLeft, 3, (255, 0, 0), -1)
-            trait_img = cv2.circle(orig, extRight, 3, (255, 0, 0), -1)
-            trait_img = cv2.circle(orig, extTop, 3, (255, 0, 0), -1)
-            trait_img = cv2.circle(orig, extBot, 3, (255, 0, 0), -1)
-
-            max_width = dist.euclidean(extLeft, extRight)
-            max_height = dist.euclidean(extTop, extBot)
-
-            if max_width > max_height:
-                trait_img = cv2.line(orig, extLeft, extRight, (0, 255, 0), 2)
-            else:
-                trait_img = cv2.line(orig, extTop, extBot, (0, 255, 0), 2)
-
-            print("Width and height are {0:.2f},{1:.2f}... \n".format(w, h))
-
-    return trait_img, area, solidity, w, h
-
-
-def RGB2HEX(color):
-    return "#{:02x}{:02x}{:02x}".format(int(color[0]), int(color[1]), int(color[2]))
-
-
-class ComputeCurvature:
-
-    def __init__(self,x,y):
-        """ Initialize some variables """
-        self.xc = 0  # X-coordinate of circle center
-        self.yc = 0  # Y-coordinate of circle center
-        self.r = 0   # Radius of the circle
-        self.xx = np.array([])  # Data points
-        self.yy = np.array([])  # Data points
-        self.x = x  # X-coordinate of circle center
-        self.y = y  # Y-coordinate of circle center
-
-    def calc_r(self, xc, yc):
-        """ calculate the distance of each 2D points from the center (xc, yc) """
-        return np.sqrt((self.xx-xc)**2 + (self.yy-yc)**2)
-
-    def f(self, c):
-        """ calculate the algebraic distance between the data points and the mean circle centered at c=(xc, yc) """
-        ri = self.calc_r(*c)
-        return ri - ri.mean()
-
-    def df(self, c):
-        """ Jacobian of f_2b
-        The axis corresponding to derivatives must be coherent with the col_deriv option of leastsq"""
-        xc, yc = c
-        df_dc = np.empty((len(c), self.x.size))
-
-        ri = self.calc_r(xc, yc)
-        df_dc[0] = (xc - self.x)/ri                   # dR/dxc
-        df_dc[1] = (yc - self.y)/ri                   # dR/dyc
-        df_dc = df_dc - df_dc.mean(axis=1)[:, np.newaxis]
-        return df_dc
-
-    def fit(self, xx, yy):
-        self.xx = xx
-        self.yy = yy
-        center_estimate = np.r_[np.mean(xx), np.mean(yy)]
-        center = optimize.leastsq(self.f, center_estimate, Dfun=self.df, col_deriv=True)[0]
-
-        self.xc, self.yc = center
-        ri = self.calc_r(*center)
-        self.r = ri.mean()
-
-        return 1 / self.r  # Return the curvature
-
-
-def skeleton_bw(thresh):
-    # Convert mask to boolean image, rather than 0 and 255 for skimage to use it
-
-    # convert an image from OpenCV to skimage
-    thresh_sk = img_as_float(thresh)
-
-    image_bw = img_as_bool((thresh_sk))
-
-    # skeleton = morphology.skeletonize(image_bw)
-
-    skeleton = morphology.thin(image_bw)
-
-    skeleton_img = skeleton.astype(np.uint8) * 255
-
-    return skeleton_img, skeleton
-
-
-def get_cmap(n, name = 'hsv'):
-    '''Returns a function that maps each index in 0, 1, ..., n-1 to a distinct
-    RGB color; the keyword argument name must be a standard mpl colormap name.'''
-    return plt.cm.get_cmap(name, n)
-
-
-def leaf_traits_computation(orig, labels, output_directory):
-    gray = cv2.cvtColor(orig, cv2.COLOR_BGR2GRAY)
-
-    leaf_index_rec = []
-    contours_rec = []
-    area_rec = []
-    curv_rec = []
-    solidity_rec = []
-    major_axis_rec = []
-    minor_axis_rec = []
-
-    leaf_color_ratio_rec = []
-    leaf_color_value_rec = []
-
-    box_coord_rec = []
-
-    num_clusters = 5
-
-    # curvature computation
-    # loop over the unique labels returned by the Watershed algorithm
-    for index, label in enumerate(np.unique(labels), start=1):
-        # if the label is zero, we are examining the 'background'
-        # so simply ignore it
-        if label == 0: continue
-
-        # otherwise, allocate memory for the label region and draw
-        # it on the mask
-        mask = np.zeros(gray.shape, dtype="uint8")
-        mask[labels == label] = 255
-
-        # get the medial axis of the contour
-        image_skeleton, skeleton = skeleton_bw(mask)
-
-        # apply individual object mask
-        masked = cv2.bitwise_and(orig, orig, mask=mask)
-
-        # individual leaf segmentation and color analysis
-        ################################################################################
-
-
-        # define result path
-        cv2.imwrite(join(output_directory, 'leaf_' + str(label) + '.png'), masked)
-
-        # save _skeleton result
-        cv2.imwrite(join(output_directory, 'leaf_skeleton_' + str(label) + '.png'), img_as_ubyte(image_skeleton))
-
-        # save color quantization result
-        # rgb_colors = color_quantization(image, thresh, save_path, num_clusters)
-        (rgb_colors, counts, hex_colors) = color_region(masked, mask, output_directory, num_clusters, False)
-
-        list_counts = list(counts.values())
-        color_ratio = []
-
-        for value_counts, value_hex in zip(list_counts, hex_colors):
-            color_ratio.append(percentage(value_counts, np.sum(list_counts)))
-
-        leaf_color_ratio_rec.append(color_ratio)
-        leaf_color_value_rec.append(hex_colors)
-
-        # detect contours in the mask and grab the largest one
-        contours, hierarchy = cv2.findContours(mask.copy(), cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-        c = max(contours, key=cv2.contourArea)
-
-        if len(c) >= 10:
-
-            contours_rec.append(c)
-            area_rec.append(cv2.contourArea(c))
-
-        else:
-            # optional to "delete" the small contours
-            # label_trait = cv2.drawContours(orig, [c], -1, (0, 0, 255), 2)
-            print("lack of enough points to fit ellipse")
-
-    contours_rec_sorted = [x for _, x in sorted(zip(area_rec, contours_rec), key=lambda pair: pair[0])]
-
-    cmap = get_cmap(len(contours_rec_sorted) + 1)
-    tracking_backgd = np.zeros(gray.shape, dtype="uint8")
-
-    # clean area record list
-    area_rec = []
-    # individual leaf traits sorting based on area order
-    ################################################################################
-    for i in range(len(contours_rec_sorted)):
-
-        c = contours_rec_sorted[i]
-
-        # assign unique color value in opencv format
-        color_rgb = tuple(reversed(cmap(i)[:len(cmap(i)) - 1]))
-
-        color_rgb = tuple([255 * x for x in color_rgb])
-
-        rect = cv2.minAreaRect(c)
-        box = cv2.boxPoints(rect)
-        box = np.array(box, dtype="int")
-        box_coord_flat = box.flatten()
-
-        box_coord = []
-        for item in box_coord_flat:
-            box_coord.append(item)
-
-        # draw a circle enclosing the object
-        ((x, y), r) = cv2.minEnclosingCircle(c)
-        # label_trait = cv2.circle(orig, (int(x), int(y)), 3, (0, 255, 0), 2)
-
-        label_trait = cv2.drawContours(orig, [c], -1, color_rgb, 2)
-
-        label_trait = cv2.putText(orig, "#{}".format(i + 1), (int(x) - 10, int(y)), cv2.FONT_HERSHEY_SIMPLEX, 0.6, color_rgb, 1)
-        # label_trait = cv2.putText(backgd, "#{}".format(i+1), (int(x) - 10, int(y)), cv2.FONT_HERSHEY_SIMPLEX, 0.6, color_rgb, 1)
-
-        #######################################individual leaf curvature computation
-
-        # Get rotated bounding ellipse of contour
-        ellipse = cv2.fitEllipse(c)
-
-        # get paramters of ellipse
-        ((xc, yc), (d1, d2), angle) = ellipse
-        track_trait = cv2.circle(tracking_backgd, (int(xc), int(yc)), 5, (255, 255, 255), -1)
-
-        if angle > 90: angle = angle - 90
-        else: angle = angle + 90
-
-        # draw major radius
-        rmajor = max(d1, d2) / 2
-        rminor = min(d1, d2) / 2
-
-        xtop = xc + math.cos(math.radians(angle)) * rmajor
-        ytop = yc + math.sin(math.radians(angle)) * rmajor
-        xbot = xc + math.cos(math.radians(angle + 180)) * rmajor
-        ybot = yc + math.sin(math.radians(angle + 180)) * rmajor
-
-        label_trait = cv2.line(orig, (int(xtop), int(ytop)), (int(xbot), int(ybot)), color_rgb, 1)
-
-        c_np = np.vstack(c).squeeze()
-        x = c_np[:, 0]
-        y = c_np[:, 1]
-
-        comp_curv = ComputeCurvature(x, y)
-        curvature = comp_curv.fit(x, y)
-
-        # compute solidity
-        solidity = float(cv2.contourArea(c)) / cv2.contourArea(cv2.convexHull(c))
-
-        # record all traits
-        leaf_index_rec.append(i)
-        area_rec.append(cv2.contourArea(c))
-        curv_rec.append(curvature)
-
-        solidity_rec.append(solidity)
-        major_axis_rec.append(rmajor)
-        minor_axis_rec.append(rminor)
-
-        box_coord_rec.append(box_coord)
-
-    n_contours = len(contours_rec_sorted)
-
-    if n_contours > 0: print('average curvature = {0:.2f}\n'.format(sum(curv_rec) / n_contours))
-    else: n_contours = 1.0
-
-    return sum(
-        curv_rec) / n_contours, label_trait, track_trait, leaf_index_rec, contours_rec, area_rec, curv_rec, solidity_rec, major_axis_rec, minor_axis_rec, leaf_color_ratio_rec, leaf_color_value_rec, box_coord_rec
-
-
-def percentage(part, whole):
-    # percentage = "{:.0%}".format(float(part)/float(whole))
-
-    percentage = "{:.2f}".format(float(part) / float(whole))
-
-    return str(percentage)
-
-
-def isbright(image_file):
-    # Set up threshold value for luminous channel, can be adjusted and generalized
-    thresh = 1.5
-
-    # Load image file
-    orig = cv2.imread(image_file)
-
-    # Make backup image
-    image = orig.copy()
-
-    # Get file name
-    # abs_path = os.path.abspath(image_file)
-
-    # filename, file_extension = os.path.splitext(abs_path)
-    # base_name = os.path.splitext(os.path.basename(filename))[0]
-
-    image_file_name = Path(image_file).name
-
-    # Convert color space to LAB format and extract L channel
-    L, A, B = cv2.split(cv2.cvtColor(image, cv2.COLOR_BGR2LAB))
-
-    # Normalize L channel by dividing all pixel values with maximum pixel value
-    L = L / np.max(L)
-
-    text_bool = "bright" if np.mean(L) < thresh else "dark"
-
-    # return image_file_name, np.mean(L), text_bool
-
-    print("np.mean(L) < thresh = {}".format(np.mean(L)))
-
-    return np.mean(L) < thresh
-
-
-def watershed_seg(orig, thresh, min_distance_value):
-    # compute the exact Euclidean distance from every binary
-    # pixel to the nearest zero pixel, then find peaks in this
-    # distance map
-    D = ndimage.distance_transform_edt(thresh)
-
-    localMax = peak_local_max(D, indices=False, min_distance=min_distance_value, labels=thresh)
-
-    # perform a connected component analysis on the local peaks,
-    # using 8-connectivity, then appy the Watershed algorithm
-    markers = ndimage.label(localMax, structure=np.ones((3, 3)))[0]
-
-    # print("markers")
-    # print(type(markers))
-
-    labels = watershed(-D, markers, mask=thresh)
-
-    print("[INFO] {} unique segments found\n".format(len(np.unique(labels)) - 1))
-
-    return labels
-
-
-def color_region(image, mask, save_path, num_clusters, write=True):
-    # read the image
-    # grab image width and height
-    (h, w) = image.shape[:2]
-
-    # apply the mask to get the segmentation of plant
-    masked_image_ori = cv2.bitwise_and(image, image, mask=mask)
-
-    # define result path for labeled images
-    if write: cv2.imwrite(join(save_path, 'masked.png'), masked_image_ori)
-
-    # convert to RGB
-    image_RGB = cv2.cvtColor(masked_image_ori, cv2.COLOR_BGR2RGB)
-
-    # reshape the image to a 2D array of pixels and 3 color values (RGB)
-    pixel_values = image_RGB.reshape((-1, 3))
-
-    # convert to float
-    pixel_values = np.float32(pixel_values)
-
-    # define stopping criteria
-    criteria = (cv2.TERM_CRITERIA_EPS + cv2.TERM_CRITERIA_MAX_ITER, 100, 0.2)
-
-    # number of clusters (K)
-    # num_clusters = 5
-    compactness, labels, (centers) = cv2.kmeans(pixel_values, num_clusters, None, criteria, 10, cv2.KMEANS_RANDOM_CENTERS)
-
-    # convert back to 8 bit values
-    centers = np.uint8(centers)
-
-    # flatten the labels array
-    labels_flat = labels.flatten()
-
-    # convert all pixels to the color of the centroids
-    segmented_image = centers[labels_flat]
-
-    # reshape back to the original image dimension
-    segmented_image = segmented_image.reshape(image_RGB.shape)
-
-    segmented_image_BRG = cv2.cvtColor(segmented_image, cv2.COLOR_RGB2BGR)
-    # define result path for labeled images
-    if write: cv2.imwrite(join(save_path, 'clustered.png'), segmented_image_BRG)
-
-    '''
-    fig = plt.figure()
-    ax = Axes3D(fig)        
-    for label, pix in zip(labels, segmented_image):
-        ax.scatter(pix[0], pix[1], pix[2], color = (centers))
-
-    result_file = (save_path + base_name + 'color_cluster_distributation.png')
-    plt.savefig(result_file)
-    '''
-    # Show only one chosen cluster
-    # masked_image = np.copy(image)
-    masked_image = np.zeros_like(image_RGB)
-
-    # convert to the shape of a vector of pixel values
-    masked_image = masked_image.reshape((-1, 3))
-    # color (i.e cluster) to render
-    # cluster = 2
-
-    cmap = get_cmap(num_clusters + 1)
-
-    # clrs = sns.color_palette('husl', n_colors = num_clusters)  # a list of RGB tuples
-
-    color_conversion = interp1d([0, 1], [0, 255])
-
-    for cluster in range(num_clusters):
-
-        print("Processing color cluster {0} ...\n".format(cluster))
-        # print(clrs[cluster])
-        # print(color_conversion(clrs[cluster]))
-
-        masked_image[labels_flat == cluster] = centers[cluster]
-
-        # print(centers[cluster])
-
-        # convert back to original shape
-        masked_image_rp = masked_image.reshape(image_RGB.shape)
-
-        # masked_image_BRG = cv2.cvtColor(masked_image, cv2.COLOR_RGB2BGR)
-        # cv2.imwrite('maksed.png', masked_image_BRG)
-
-        gray = cv2.cvtColor(masked_image_rp, cv2.COLOR_BGR2GRAY)
-
-        # threshold the image, then perform a series of erosions +
-        # dilations to remove any small regions of noise
-        thresh = cv2.threshold(gray, 0, 255, cv2.THRESH_BINARY)[1]
-
-        # thresh_cleaned = clear_border(thresh)
-
-        cnts = cv2.findContours(thresh.copy(), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-        cnts = imutils.grab_contours(cnts)
-        # c = max(cnts, key=cv2.contourArea)
-
-        '''
-        # compute the center of the contour area and draw a circle representing the center
-        M = cv2.moments(c)
-        cX = int(M["m10"] / M["m00"])
-        cY = int(M["m01"] / M["m00"])
-        # draw the countour number on the image
-        result = cv2.putText(masked_image_rp, "#{}".format(cluster + 1), (cX - 20, cY), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (255, 255, 255), 2)
-        '''
-
-        if not cnts:
-            print("findContours is empty")
-        else:
-
-            # loop over the (unsorted) contours and draw them
-            for (i, c) in enumerate(cnts):
-                result = cv2.drawContours(masked_image_rp, c, -1, color_conversion(np.random.random(3)), 2)
-                # result = cv2.drawContours(masked_image_rp, c, -1, color_conversion(clrs[cluster]), 2)
-
-            # result = result(np.where(result == 0)== 255)
-            result[result == 0] = 255
-
-            result_BRG = cv2.cvtColor(result, cv2.COLOR_RGB2BGR)
-            cv2.imwrite(join(save_path, 'result_' + str(cluster) + '.png'), result_BRG)
-
-    counts = Counter(labels_flat)
-    # sort to ensure correct color percentage
-    counts = dict(sorted(counts.items()))
-
-    center_colors = centers
-
-    # We get ordered colors by iterating through the keys
-    ordered_colors = [center_colors[i] for i in counts.keys()]
-    hex_colors = [RGB2HEX(ordered_colors[i]) for i in counts.keys()]
-    rgb_colors = [ordered_colors[i] for i in counts.keys()]
-
-    # print(hex_colors)
-
-    index_bkg = [index for index in range(len(hex_colors)) if hex_colors[index] == '#000000']
-
-    # print(index_bkg[0])
-
-    # print(counts)
-    # remove background color
-    del hex_colors[index_bkg[0]]
-    del rgb_colors[index_bkg[0]]
-
-    # Using dictionary comprehension to find list
-    # keys having value .
-    delete = [key for key in counts if key == index_bkg[0]]
-
-    # delete the key
-    for key in delete: del counts[key]
-
-    fig = plt.figure(figsize=(6, 6))
-    plt.pie(counts.values(), labels=hex_colors, colors=hex_colors)
-
-    # define result path for labeled images
-    plt.savefig(join(save_path, 'pie_color.png'))
-
-    return rgb_colors, counts, hex_colors
-
-
-def trait_extract(image, output_directory):
-    # gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-
-    abs_path = abspath(image)
-    filename, _ = splitext(abs_path)
-    base_name = splitext(basename(filename))[0]
-    file_size = getsize(image) / float(1 << 20)
-    image_file_name = Path(image).name
-
-    # make the folder to store the results
-    # current_path = abs_path + '/'
-
-    print("Exacting traits for image : {0}\n".format(str(image_file_name)))
-
-    print("Results_folder: " + output_directory)
-
-    if isbright(image):
-        if file_size > 5.0: print("It will take some time due to larger file size {0} MB".format(str(int(file_size))))
-
-        image = cv2.imread(image)
-        image_copy = image.copy()
-
-        # color clustering based plant object segmentation
-        thresh = color_cluster_seg(image_copy, 'lab', '1', 2, 200)
-        cv2.imwrite(join(output_directory, base_name + '_seg.png'), thresh)
-
-        # find external contour
-        (trait_img, area, solidity, max_width, max_height) = comp_external_contour_2(image.copy(), thresh)
-        cv2.imwrite(join(output_directory, base_name + '_excontour.png'), trait_img)
-
-        num_clusters = 5
-        (rgb_colors, counts, hex_colors) = color_region(image_copy, thresh, output_directory, num_clusters)
-
-        list_counts = list(counts.values())
-        color_ratio = []
-
-        for value_counts, value_hex in zip(list_counts, hex_colors):
-            color_ratio.append(percentage(value_counts, np.sum(list_counts)))
-
-        ############################################## leaf number computation
-
-        if area > 20000:
-            min_distance_value = 10
-        elif area > 10000:
-            min_distance_value = 7
-        elif area > 5000:
-            min_distance_value = 7
-        else:
-            min_distance_value = 5
-
-        print("min_distance_value = {}".format(min_distance_value))
-
-        # watershed based leaf area segmentaiton
-        labels = watershed_seg(image_copy, thresh, min_distance_value)
-
-        # save watershed result label image
-        # Map component labels to hue val
-        label_hue = np.uint8(128 * labels / np.max(labels))
-        # label_hue[labels == largest_label] = np.uint8(15)
-        blank_ch = 255 * np.ones_like(label_hue)
-        labeled_img = cv2.merge([label_hue, blank_ch, blank_ch])
-
-        # cvt to BGR for display
-        labeled_img = cv2.cvtColor(labeled_img, cv2.COLOR_HSV2BGR)
-
-        # set background label to black
-        labeled_img[label_hue == 0] = 0
-        cv2.imwrite(join(output_directory, base_name + '_label.png'), labeled_img)
-
-        (avg_curv, label_trait, track_trait, leaf_index_rec, contours_rec, area_rec, curv_rec, solidity_rec, major_axis_rec, minor_axis_rec,
-         leaf_color_ratio_rec, leaf_color_value_rec, box_coord_rec) = leaf_traits_computation(image_copy, labels, output_directory)
-
-        wb = openpyxl.Workbook()
-        sheet = wb.active
-
-        sheet_leaf = wb.create_sheet()
-
-        sheet.cell(row=1, column=1).value = 'c1x'
-        sheet.cell(row=1, column=2).value = 'c1y'
-        sheet.cell(row=1, column=3).value = 'c2x'
-        sheet.cell(row=1, column=4).value = 'c2y'
-        sheet.cell(row=1, column=5).value = 'c3x'
-        sheet.cell(row=1, column=6).value = 'c3y'
-        sheet.cell(row=1, column=7).value = 'c4x'
-        sheet.cell(row=1, column=8).value = 'c4y'
-
-        for row in box_coord_rec: sheet.append(row)
-
-        excel_file_path = join(output_directory, 'bbox.xlsx')
-        wb.save(excel_file_path)
-        wb = openpyxl.load_workbook(excel_file_path)
-        sh = wb.active  # was .get_active_sheet()
-
-        csv_file_path = join(output_directory, 'bbox.csv')
-        with open(csv_file_path, 'w', newline="") as f:
-            c = csv.writer(f)
-            for r in sh.rows:  # generator; was sh.rows
-                c.writerow([cell.value for cell in r])
-
-        n_leaves = int(len(leaf_index_rec))
-
-        # save watershed result label image
-        result_file = join(output_directory, base_name + '_leafspec.png')
-        cv2.imwrite(result_file, label_trait)
-
-        # save watershed result label image
-        result_file = join(output_directory, base_name + '_trace.png')
-        cv2.imwrite(result_file, track_trait)
-    else:
-        area = solidity = max_width = max_height = avg_curv = n_leaves = 0
-
-    return image_file_name, area, solidity, max_width, max_height, avg_curv, n_leaves, color_ratio, hex_colors, leaf_index_rec, area_rec, curv_rec, solidity_rec, major_axis_rec, minor_axis_rec, leaf_color_ratio_rec, leaf_color_value_rec
-
-
-@cli.command()
-@click.argument('source')
-@click.option('-o', '--output_directory', required=False, type=str, default='')
-@click.option('-ft', '--file_types', required=False, type=str, default='jpg,png')
-@click.option('-t', '--template', required=False, type=str, default='/opt/smart/marker_template/sticker_template.png')
-@click.option('-m', '--multiprocessing', is_flag=True)
-def traits(source, output_directory, file_types, template, multiprocessing):
-    Path(output_directory).mkdir(parents=True, exist_ok=True)
-
-    if Path(source).is_file():
-        segmentation(source, output_directory)
-        trait_extract(source, output_directory)
-    # elif Path(source).is_dir():
-    #     result_list = []
-    #     result_list_leaf = []
-
-    #     processes = cpu_count()
-    #     print(f"Using up to {processes} processes to extract traits from {len(files)} images")
-    #     with closing(Pool(processes=processes)) as pool:
-    #         seg_results = pool.map(trait_extract, images)
-    #         traits_results = pool.map(trait_extract, images)
-    #         pool.terminate()
-
-
-# @cli.command()
-# @click.argument('source')
-# @click.option('-o', '--output_directory', required=False, type=str, default='')
-# @click.option('-ft', '--file_types', required=False, type=str, default='jpg,png')
-# @click.option('-r', '--replace', is_flag=True)
-# def enhance(source, output_directory, file_types, replace):
-#     Path(output_directory).mkdir(parents=True, exist_ok=True)
-#
-#     get_path = lambda i: f"{join(i.output_directory, i.input_stem)}.png" if replace else f"{join(i.output_directory, i.input_stem)}.enhanced.png"
-#
-#     if Path(source).is_file():
-#         input = ImageInput(input_file=source, output_directory=output_directory)
-#         enhanced = image_enhance(cv2.imread(input.input_file))
-#         enhanced.save(get_path(input))
-#
-#     elif Path(source).is_dir():
-#         patterns = [ft.lower() for ft in file_types.split(',')]
-#         if 'jpg' in patterns:
-#             patterns.append('jpeg')
-#         if len(patterns) == 0:
-#             raise ValueError(f"You must specify file types!")
-#
-#         patterns = patterns + [pattern.upper() for pattern in patterns]
-#         files = sum((sorted(glob(join(source, f"*.{file_type}"))) for file_type in patterns), [])
-#         inputs = [ImageInput(input_file=file, output_directory=output_directory) for file in files]
-#         print(f"Found {len(files)} files with extensions {', '.join(patterns)}: \n" + '\n'.join(files))
-#
-#         for input in inputs:
-#             # img_hsv = cv2.cvtColor(cropped, cv2.COLOR_BGR2HSV)
-#             # img_hsv[:, :, 2] = cv2.equalizeHist(img_hsv[:, :, 2])
-#             # enhanced = cv2.cvtColor(img_hsv, cv2.COLOR_HSV2BGR)
-#
-#             img_hsv = cv2.cvtColor(cv2.imread(input.input_file), cv2.COLOR_BGR2YUV)
-#             img_hsv[:, :, 0] = cv2.equalizeHist(img_hsv[:, :, 0])
-#             enhanced = cv2.cvtColor(img_hsv, cv2.COLOR_YUV2BGR)
-#
-#             cv2.imwrite(get_path(input), enhanced)
-#
-#             # enhanced = image_enhance(cropped)
-#             # enhanced.save(f"{join(image.output_directory, image.input_stem)}.png")
-#     else:
-#         print(f"Path does not exist: {source}")
-
-
-# @cli.command()
-# @click.argument('source')
-# @click.option('-o', '--output_directory', required=False, type=str, default='')
-# @click.option('-ft', '--file_types', required=False, type=str, default='jpg,png')
-# @click.option('-l', '--luminosity_threshold', required=False, type=float, default=0.1)
-# @click.option('-t', '--template', required=False, type=str, default='marker_template.png')
-# @click.option('-m', '--multiprocessing', is_flag=True)
-# def traits(source, output_directory, file_types, luminosity_threshold, template, multiprocessing):
-#     Path(output_directory).mkdir(parents=True, exist_ok=True)
-#
-#     if Path(source).is_file():
-#         # check luminosity
-#         image = ImageInput(input_file=source, output_directory=output_directory)
-#         print(f"Checking image quality")
-#         if check_discard_merge2([image], luminosity_threshold):
-#             print(f"{source} is too dark!")
-#         else:
-#             print(f"{source} is light enough.")
-#
-#         # crop
-#         input = ImageInput(input_file=join(output_directory, Path(source).name), output_directory=output_directory)
-#         cropped = circle_detect(input.input_file, template)
-#         enhanced = image_enhance(cropped)
-#         enhanced.save(f"{join(input.output_directory, input.input_stem)}.png")
-#
-#         # extract traits
-#         image = ImageInput(input_file=join(output_directory, Path(source).name), output_directory=output_directory)
-#         result = trait_extract(image)
-#         write_results(image.output_directory, [result])
-#     elif Path(source).is_dir():
-#         patterns = [ft.lower() for ft in file_types.split(',')]
-#         if 'jpg' in patterns: patterns.append('jpeg')
-#         if len(patterns) == 0: raise ValueError(f"You must specify file types!")
-#
-#         patterns = patterns + [pattern.upper() for pattern in patterns]
-#         files = sum((sorted(glob(join(source, f"*.{file_type}"))) for file_type in patterns), [])
-#         images = [ImageInput(input_file=file, output_directory=output_directory) for file in files]
-#
-#         print(f"Found {len(files)} files with extensions {', '.join(patterns)}: \n" + '\n'.join(files))
-#
-#         # check luminosity
-#         check_discard_merge2(images, luminosity_threshold)
-#
-#         # manipulate images in the output directory from here on
-#         images = [ImageInput(input_file=join(output_directory, Path(file).name), output_directory=output_directory) for file in files]
-#
-#         # crop
-#         for input in images:
-#             cropped = circle_detect(input.input_file, template)
-#             if cropped.size == 0:
-#                 print(f"No circle found, nothing to crop")
-#             else:
-#                 cv2.imwrite(f"{join(input.output_directory, input.input_stem)}.png", cropped)
-#
-#         # extract traits
-#         if multiprocessing:
-#             processes = cpu_count()
-#             print(f"Using up to {processes} processes to extract traits from {len(files)} images")
-#             with closing(Pool(processes=processes)) as pool:
-#                 results = pool.map(trait_extract, images)
-#                 pool.terminate()
-#         else:
-#             print(f"Using a single process to extract traits from {len(files)} images")
-#             results = [trait_extract(image) for image in images]
-#
-#         write_results(images[0].output_directory, results)
-#     else:
-#         print(f"File not found: {source}")
-
-
-if __name__ == '__main__':
-    cli()
