+++ core/demo_color_seg.py
