+++ core/histogram_equalization.py
