+++ plantit_pipeline.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+python3 /opt/smart/core/color_seg.py -p $INPUT -ft $FILETYPES
+python3 /opt/smart/core/trait_extract_parallel_ori.py -p $INPUT -ft $FILETYPES
